refactor: enhance type safety by using static_cast for uint8_t conversions in Board and related classes
refactor: remove CLI executable and associated code from Makefile and project structure

refactor: optimize rendering calculations in GameBoardRenderer and GameWindow for better performance

+++ include/gomoku/core/Board.hpp
@@ -58,7 +58,7 @@ public:
 
 private:
     static constexpr int N = BOARD_SIZE * BOARD_SIZE;
-    static constexpr uint16_t idx(uint8_t x, uint8_t y) { return y * BOARD_SIZE + x; }
+    static constexpr uint16_t idx(uint8_t x, uint8_t y) { return static_cast<uint16_t>(y * BOARD_SIZE + x); }
 
     std::array<Cell, N> cells {};
     Player currentPlayer { Player::Black };
